✨ feat: added lazy loading and remove additional layouts
