+++ src/components/auth/VerifyCodeForm.tsx
@@ -1,0 +1,106 @@
+"use client";
+import React from "react";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { SubmitHandler, useForm } from "react-hook-form";
+import { useRouter } from "next/navigation";
+import * as z from "zod";
+import { REGEXP_ONLY_DIGITS } from "input-otp";
+import axios, { AxiosError } from "axios";
+import { toast } from "sonner";
+
+import { verifySchema } from "@/schemas/verifySchema";
+import {
+  Form,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormControl,
+  FormMessage,
+} from "@/components/ui/form";
+import {
+  InputOTP,
+  InputOTPGroup,
+  InputOTPSlot,
+} from "@/components/ui/input-otp";
+import { Button } from "@/components/ui/button";
+import { CardContent } from "@/components/ui/card";
+import { ApiResponse } from "@/types/ApiResponse";
+
+type VerifyCodeFormProps = {
+  username: string;
+};
+
+const VerifyCodeForm = ({
+  username,
+}: VerifyCodeFormProps): React.JSX.Element => {
+  const router = useRouter();
+
+  const onSubmit: SubmitHandler<z.infer<typeof verifySchema>> = async (
+    _data
+  ) => {
+    try {
+      const res = await axios.post<ApiResponse>("/api/verify-code", {
+        username,
+        code: _data.code,
+      });
+
+      toast(res.data.message);
+      router.replace("/sign-in");
+    } catch (error) {
+      const err = error as AxiosError<ApiResponse>;
+      toast(err.response?.data.message);
+    }
+  };
+
+  const form = useForm<z.infer<typeof verifySchema>>({
+    resolver: zodResolver(verifySchema),
+    defaultValues: {
+      code: "",
+    },
+    mode: "onSubmit",
+  });
+
+  return (
+    <Form {...form}>
+      <form onSubmit={form.handleSubmit(onSubmit)} className="w-full space-y-6">
+        <CardContent className="flex flex-col gap-4">
+          <FormField
+            control={form.control}
+            name="code"
+            render={({ field }) => (
+              <FormItem className="">
+                <FormLabel>One-Time Password</FormLabel>
+                <FormControl>
+                  <InputOTP
+                    pattern={REGEXP_ONLY_DIGITS}
+                    maxLength={6}
+                    {...field}
+                  >
+                    <InputOTPGroup>
+                      <InputOTPSlot index={0} />
+                      <InputOTPSlot index={1} />
+                      <InputOTPSlot index={2} />
+                      <InputOTPSlot index={3} />
+                      <InputOTPSlot index={4} />
+                      <InputOTPSlot index={5} />
+                    </InputOTPGroup>
+                  </InputOTP>
+                </FormControl>
+
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <Button
+            disabled={!form.formState.isValid || form.formState.isSubmitting}
+            type="submit"
+          >
+            {form.formState.isSubmitting ? "Checking your OTP..." : "Submit"}
+          </Button>
+        </CardContent>
+      </form>
+    </Form>
+  );
+};
+
+export default VerifyCodeForm;
